Fix issue with nobkg fitting and merge nsigma and eff plotting macros
